moves the re-usable code into conftest.py [#164865987]
